+++ src/main/kotlin/com/github/israelkli/intellijplugincopyfilewithproblems/actions/CopyFileWithInlineIssues.kt
@@ -55,6 +55,5 @@ class CopyFileWithInlineIssues : BaseFileAction() {
                        (!virtualFile.isDirectory)
 
         e.presentation.isEnabledAndVisible = isEnabled
-        e.presentation.text = "Copy File with Inline Issues"
     }
 }