+++ src/main/kotlin/com/github/israelkli/intellijplugincopyfilewithproblems/actions/CopyWithInlineIssues.kt
@@ -61,6 +61,5 @@ class CopyWithInlineIssues : BaseFileAction() {
                        (editor.selectionModel.hasSelection())
 
         e.presentation.isEnabledAndVisible = isEnabled
-        e.presentation.text = "Copy with Inline Issues"
     }
 }